+++ pkg/datafetcher/file_fetcher.go
@@ -5,10 +5,10 @@ import (
 	"os"
 )
 
-// FileFetcher fetches data from a file.
-type FileFetcher struct{}
+// fileFetcher fetches data from a file.
+type fileFetcher struct{}
 
-func (f FileFetcher) FetchData(source string) ([]byte, error) {
+func (f fileFetcher) FetchData(source string) ([]byte, error) {
 	data, err := os.ReadFile(source)
 	if err != nil {
 		return nil, fmt.Errorf("failed to read file: %w", err)
